requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
django-model-utils==3.2.0
